2.1.0
Finally, restrictions for all commands has been implemented. This is an all or nothing deal, since the bot really only does a few things.
There is also now a global override for admin perms.
there is an option to restrict the send channel, however it is not implemented yet, planned for 2.1.1.